RavenDB-27222 Address PR comments: merge the pg catalog tests and share the pg test helpers
